backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
openpyxl==3.1.2
python-multipart==0.0.9
google-generativeai==0.3.2
pydantic-settings==2.1.0
slowapi==0.1.9
emails==0.6
